% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\docType{class}
\name{AnnotationParam-class}
\alias{AnnotationParam-class}
\title{A class containing the annotation parameters}
\description{
A class containing the annotation parameters
}
\section{Slots}{

\describe{
\item{\code{da_tol}}{`numeric(1)` m/z tolerance in Dalton}

\item{\code{rt_tol}}{`numeric(1)` rT tolerance in sec}

\item{\code{abd_tol}}{`numeric(1)` relative abundance tolerance, each peak which
have an higher difference of relative abundance with its corresponding
theoretical peak will be discarded}

\item{\code{adduct_names}}{`character vector` adduct names from the enviPat package}

\item{\code{instrument}}{`character(1)` instrument names from the enviPat package}

\item{\code{database}}{`character(1)` name of the database to load}

\item{\code{cpd_classes}}{`character vector` compound classes in database to
restrict for annotation}
}}

