% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{AnnotationParam}
\alias{AnnotationParam}
\title{AnnotationParam}
\usage{
AnnotationParam(
  da_tol = 0.015,
  rt_tol = 10,
  abd_tol = 25,
  adduct_names = c("[M+Na]+", "[M+NH4]+", "[M+H-H2O]+", "[M+H]+", "[M-H]-"),
  instrument = "QTOF_XevoG2-S_R25000@200"
)
}
\arguments{
\item{da_tol}{`numeric(1)` m/z tolerance in Dalton}

\item{rt_tol}{`numeric(1)` rT tolerance in sec}

\item{abd_tol}{`numeric(1)` relative abundance tolerance, each peak which
have an higher difference of relative abundance with its corresponding
theoretical peak will be discarded}

\item{adduct_names}{`character vector` adduct names from the enviPat package}

\item{instrument}{`character(1)` instrument names from the enviPat package}
}
\value{
`AnnotationParam` object
}
\description{
Create an AnnotationParam object
}
