% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{check_ms_file}
\alias{check_ms_file}
\title{Check mass spectrometry file}
\usage{
check_ms_file(filepath, exp_polarity)
}
\arguments{
\item{exp_polarity}{"positive" or "negative" only !}
}
\value{
xcmsRaw object
}
\description{
check if a mass spectrometry file can be read by xcms
     & if the polarity expected is the good one
}
