% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{check_ms_process_args}
\alias{check_ms_process_args}
\title{Check all args of the processing workflow}
\usage{
check_ms_process_args(
  raw_files,
  sqlite_path,
  converter,
  cwt_params,
  obw_params,
  pd_params,
  ann_params,
  cores
)
}
\arguments{
\item{raw_files}{`character vector` filepaths to the raw files}

\item{sqlite_path}{`character(1)` filepath to the sqlite database used to
save the process results}

\item{converter}{`character(1)` filepath to the msconvert application}

\item{cwt_params}{`CentWaveParam`}

\item{obw_params}{`ObiwarpParam`}

\item{pd_params}{`PeakDensityParam`}

\item{ann_params}{`AnnotationParam`}

\item{cores}{`integer(1)` number of cores to use for the processing workflow
(generaly one core = one file peakpicked)}
}
\description{
Check all arguments of the processing workflow.
     Will raise an error if one argument is incorrect
}
