% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_FUN.R
\name{compare_spectras}
\alias{compare_spectras}
\title{Compare spectras}
\usage{
compare_spectras(
  q_spectra,
  l_spectras,
  da_tol = 0.05,
  abd_tol = 25,
  suffix = "theo"
)
}
\arguments{
\item{q_spectra}{dataframe with columns (at least) :
\itemize{
    \item mz m/z
    \item int or abd for intensity or relative abundance tolerance
}}

\item{l_spectras}{list of dataframe with columns (at least):
\itemize{
    \item mz m/z
    \item int or abd for intensity or relative abundance tolerance
}}

\item{da_tol}{Da tolerance}

\item{abd_tol}{abundance tolerance, at which \% tolerance a peak need to be}

\item{suffix}{which suffix to apply to
the column names of the library spectras}
}
\value{
list (for each library spectra) wich contains a list with :
     \itemize{
         \item score the isotopic score between query & library spectra
         \item deviation_mz the meanned m/z deviation
         \item npeak number of peaks matched in query spectra
         \item spectras list of dataframe consisting of
             the merge of the query spectra with the corresponding lines
             of the library spectra
     }
}
\description{
Compare query spectras against library spectras
The scoring algorithm will search each corresponding observed peak
     with theoreticals
Therefore it contains some important rules :
     \itemize{
         \item an observed peak can only correspond to ONE theoretical peak
             and vice versa
         \item the relative abundance peak must not be under a tolerance
             compared to the theoretical
             but it can be higher since a peak can hide another
         \item the A+x is not searched if the A+x-1 is not found
             (the loop search is stopped)
     }
}
