% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{compress}
\alias{compress}
\title{Compress object}
\usage{
compress(obj)
}
\arguments{
\item{obj}{any R object}
}
\value{
`blob` object
}
\description{
Compress an R object in order to put it in a sqlite file with the `blob` type
}
