% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_FUN.R
\name{convert_ppm_da}
\alias{convert_ppm_da}
\title{Convert ppm to Da}
\usage{
convert_ppm_da(ppm, mass)
}
\arguments{
\item{ppm}{`numeric(1)` ppm tolerance}

\item{mass}{`numeric(1)` mass}
}
\value{
`numeric(1)` ppm tolerance converted in Dalton
}
\description{
Convert ppm to Dalton according a mass
}
