% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_connect}
\alias{db_connect}
\title{Create sqlite file}
\usage{
db_connect(sqlite_path)
}
\arguments{
\item{sqlite_path}{`character(1)` filepath to the sqlite file to create}
}
\value{
`SQLiteConnection`
}
\description{
Create an sqlite file and optimize it via PRAGMA queries
}
\seealso{
RSQLite::dbConnect
}
