% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_get_nsamples}
\alias{db_get_nsamples}
\title{Count nsamples}
\usage{
db_get_nsamples(db)
}
\arguments{
\item{db}{`SQLiteConnection`}
}
\value{
`numeric` number of samples recorded in database
}
\description{
Count number of samples recorded in database
}
