% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_get_params}
\alias{db_get_params}
\title{Get all parameters}
\usage{
db_get_params(db)
}
\arguments{
\item{db}{`SQLiteConnection`}
}
\value{
`list` containing :
\itemize{
    \item filter `DataFrame` of one line with columns :
    \itemize{
        \item mz_range_min `numeric` m/z range min
        \item mz_range_max `numeric` m/z range max
        \item rt_range_min `numeric` rT range min
        \item rt_range_max `numeric` rT range max
    }
    \item cwt `DataFrame` of one line with columns :
    \itemize{
        \item ppm `numeric` Maximal tolerated m/z deviation in consecutive
        scans in parts per million (ppm)
        \item peakwidth_min `numeric` Expected approximate peak width min in
        chromatographic space
        \item peakwidth_max `numeric` Expected approximate peak width max in
        chromatographic space
        \item snthresh `numeric` Signal to noise ratio cutoff
        \item prefilter_step `numeric` Mass traces are only retained if they
        contain at least k peaks with intensity >= I
        \item prefilter_level `numeric` Mass traces are only retained if they
        contain at least k peaks with intensity >= I
        \item mzCenterFun `character` Name of the function to calculate the
        m/z center of the chromatographic peak
        \item integrate `integer` Integration method. If unchecked the
        descent is done on the real data, if checked peak limits are found
        through descent on the mexican hat filtered data. Method 1 is very
        accurate but prone to noise, while method 2 is more robust to noise
        but less exact
        \item mzdiff `numeric` Minimum difference in m/z for peaks with
        overlapping retention times, can be negative to allow overlap
        \item fitgauss `integer` whether or not a Gaussian should be fitted
        to each peak. This affects mostly the retention time position of the
        peak
        \item noise `numeric` Optional argument which is useful for data that
         was centroided without any intensity threshold, centroids with
         intensity < noise are omitted from ROI detection
        \item verboseColumns `integer` whether additional peak meta data
        columns should be returned, ignore
        \item firstBaselineCheck `integer` Continuous data within regions of
        interest is checked to be above the first baseline
    }
    \item obw `DataFrame` of one line with columns :
    \itemize{
        \item binSize `numeric` slice of overlapping m/z groups
        \item response `numeric` Defining the responsiveness of warping with
        response = 0 giving linear warping on start and end points and
        response = 100 warping using all bijective anchors
        \item distFun `character` Distance function to be used.
        Allowed values are :
        \itemize{
            \item cor : Pearson's correlation
            \item cor_opt : calculate only 10% diagonal band of distance
            matrix; better runtime)
            \item cov : covariance
            \item prd : product
            \item euc : Euclidian distance
        }
        \item gapInit `numeric` Defining the penalty for gap opening
        \item gapExtend `numeric` Defining the penalty for gap enlargement
        \item factorDiag `numeric` Defining the local weight applied to
        diagonal moves in the alignment
        \item factorGap `numeric` Defining the local weight for gap moves in
        the alignment
        \item localAlignment `integer` Whether a local alignment should be
        performed instead of the default global alignment
        \item initPenalty `numeric` Defining the penalty for initiating an
        alignment (for local alignment only)
    }
    \item pd `DataFrame` of one line with columns :
    \itemize{
        \item bw `numeric` retention time standard deviation (s) allowed
        \item minFraction `numeric` defining the minimum fraction of samples
        in at least one sample group in which the peaks have to be present
        to be considered as a peak group (feature)
        \item minSamples `integer` with the minimum number of samples in at
        least one sample group in which the peaks have to be detected to be
        considered a peak group (feature)
        \item binSize `numeric` slice of overlapping m/z groups
        \item maxFeatures `integer` with the maximum number of peak groups
        to be identified in a single mz slice
    }
    \item ann `DataFrame` of one line with columns :
    \itemize{
        \item da_tol `numeric` m/z tolerance in Dalton
        \item rt_tol `numeric` rT tolerance in sec
        \item abd_tol `numeric` relative abundance tolerance, each peak which
        have an higher difference of relative abundance with its
        corresponding theoretical peak will be discarded
        \item adduct_names `character` adduct names from the enviPat package
        collapsed with the character ";"
        \item instrument `character` instrument names from the enviPat
        package
        \item database `character(1)` name of the database used
        \item cpd_classes `character vector` compound classes in database to
         used to restrict the annotations, collapsed with the character ";"
    }
}
}
\description{
Get all parameters recorded in database
}
