% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_get_peaks}
\alias{db_get_peaks}
\title{Get peaks}
\usage{
db_get_peaks(db, feature_ids = NULL)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{feature_ids}{`integer vector` feature IDs}
}
\value{
`DataFrame` obtained from XCMS, it contains all peaks peak
picked, it contains the columns :
\itemize{
    \item feature_id `integer` feature ID
    \item mz `numeric` m/z
    \item mz_min `numeric` m/z born min
    \item mz_max `numeric` m/z born max
    \item rt `numeric` retention time
    \item rtmin `numeric` retention time born min
    \item rtmax `numeric` retention time born max
    \item int `numeric` area integrated
    \item intb `numeric` area integrated with baseline substracted
    \item maxo `numeric` maximum of intensity
    \item sn `numeric` signal / noise
    \item egauss `numeric` ignore
    \item mu `numeric` ignore
    \item sigma `numeric` ignore,
    \item h `numeric` ignore
    \item f `integer` ID of the ROI in the ROI list constructed by XCMS,
    no use
    \item dppm `numeric` m/z deviation of the peak in ppm
    \item scale `integer` centwave scale used for the integration
    (in scans & not in sec !!!)
    \item scpos `integer` scan position
    \item scmin `integer` scan born min before the optimization of the area
    integrated
    \item scmax `integer` scan born max before the optimization of the area
    integrated
    \item lmin `integer` scan born min after the optimization of the area
    integrated
    \item lmax `integer` scan born max after the optimization of the area
    integrated
    \item sample `character` sample name
    \item polarity `character` polarity
}
}
\description{
Get peaks from database
}
