% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_get_profile}
\alias{db_get_profile}
\title{Get profile matrix}
\usage{
db_get_profile(db, sample_name, polarity)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{sample_name}{`character(1)` sample name}

\item{polarity}{`character(1)` "positive" or "negative"}
}
\value{
a profile matrix
}
\description{
Get the profile matrix of corresponding sample in the polarity desired
}
