% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_get_spectra_infos}
\alias{db_get_spectra_infos}
\title{Get all spectra informations}
\usage{
db_get_spectra_infos(db)
}
\arguments{
\item{db}{`SQLiteConnection`}
}
\value{
`DataFrame`, each line correspond to a spectra
annotated, with the columns :
\itemize{
    \item spectra_id `integer` spectra ID
    \item score `numeric` isotopic score observed
    \item deviation_mz `numeric` m/z deviation observed
    \item npeak `integer` number of isotopologue annotated
    \item basepeak_int `numeric` area of the basepeak annotated
    \item sum_int `numeric` cumulative sum off all the area of the
    isotopologues annotated
    \item rt `numeric` retention time
}
}
\description{
Get all spectra informations in database
}
