% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_get_spectras}
\alias{db_get_spectras}
\title{Get spectras}
\usage{
db_get_spectras(db, spectra_ids = NULL)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{spectra_ids}{`integer vector` the spectra IDs}
}
\value{
`DataFrame`, each line correspond to a peak annotated with
its corresponding theoretical peak or the theoretical peak missed,
with the columns :
\itemize{
    \item spectra_id `integer` spectra ID
    \item feature_id `integer` feature ID
    \item mz `numeric` m/z
    \item int `numeric` area integrated
    \item abd `numeric` relative abundance
    \item ion_id_theo `integer` ignore
    \item mz_theo `numeric` theoretical m/z
    \item abd_theo `numeric` theoretical relative abundance
    \item iso_theo `character` theoretical isotopologue annotation
}
}
\description{
Get spectras from database
}
