% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_read_ms_file}
\alias{db_read_ms_file}
\title{Get ms file}
\usage{
db_read_ms_file(db, sample_name, polarity)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{sample_name}{`character(1)` sample name}

\item{polarity}{`character(1)` "positive" or "negative"}
}
\value{
an `xcmsRaw` object
}
\description{
Get the `xcmsRaw` of corresponding sample in the polarity desired
}
