% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_record_params}
\alias{db_record_params}
\title{Record parameters in database}
\usage{
db_record_params(
  db,
  filter_params,
  cwt_params,
  obw_params,
  pd_params,
  ann_params
)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{filter_params}{`FilterParam`}

\item{cwt_params}{`CentwaveParam`}

\item{obw_params}{`ObiwarpParam`}

\item{pd_params}{`PeakDensityParam`}

\item{ann_params}{`AnnotationParam`}
}
\description{
Record processing parameters in the database
}
