% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_record_samples}
\alias{db_record_samples}
\title{Record samples}
\usage{
db_record_samples(db, sample_names)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{sample_names}{`character vector` sample names, should be unique !!!!}
}
\description{
Create an empty table "sample" with all predefined sample names
The sample names will be the primary key of the table
It serve to be updated later with the function "import_ms_file"
}
