% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_record_xsa}
\alias{db_record_xsa}
\title{Record `xsAnnotate` in the database}
\usage{
db_record_xsa(db, xsa, sample_name)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{xsa}{`xsAnnotate`}

\item{sample_name}{`character(1)` sample name}
}
\description{
Record an `xsAnnotate`
This record is only use to debugging the `annotate_peaks` function to see why
some peaks were not annotated
}
