% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_record_xset}
\alias{db_record_xset}
\title{Record `xcmsSet` in the database}
\usage{
db_record_xset(db, xset_pos, xset_neg, sample_name)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{xset_pos}{`xcmsSet` object with only the positive result, can be NULL}

\item{xset_neg}{`xcmsSet` object with only the negative result, can be NULL}

\item{sample_name}{`character(1)` sample name}
}
\description{
Record an `xcmsSet` with the positive results and another with the negative
obtained by XCMS in the sample table
This record is only use to debugging the `annotate_peaks` function to see why
some peaks were not annotated
}
