% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{dbReadTable}
\alias{dbReadTable}
\title{Get table from sqlite file}
\usage{
dbReadTable(db, table_name, ...)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{table_name}{`character(1)` name of the table queried}

\item{...}{other parameters passed on the method}
}
\value{
`DataFrame` of the table
}
\description{
Override the default function `DBI::dbReadTable` in order to return an empty
`DataFrame` if RSQLite or DBI return the error "no such table"
}
\seealso{
`DBI::dbReadTable`
}
