% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{dbWriteTable}
\alias{dbWriteTable}
\title{Write a `DataFrame` in sqlite file}
\usage{
dbWriteTable(db, table_name, data, overwrite = FALSE, append = FALSE, ...)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{table_name}{`character(1)` name of the table}

\item{data}{`DataFrame`}

\item{overwrite}{`logical(1)` should overwrite if an existing table already
exists with that name ?}

\item{append}{`logical(1)` should append to the table if it exists}

\item{...}{other parameters passed on the method}
}
\description{
Override the default function `DBI::dbWriteTable` in order to force the
write process even if the database is locked by another process
}
\seealso{
`DBI::dbWriteTable`
}
