% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{decompress}
\alias{decompress}
\title{Decompress object}
\usage{
decompress(obj)
}
\arguments{
\item{obj}{`blob` object}
}
\value{
the initial R object
}
\description{
Decompress a `blob` object to obtain the initial R object
}
