% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{escape_regex}
\alias{escape_regex}
\title{Escape special characters}
\usage{
escape_regex(string)
}
\arguments{
\item{string}{`character`}
}
\value{
`character`
}
\description{
Escape all special characters
special characters are : .|()^{}+$*?[]
}
\examples{
\dontrun{escapeRegex("C:/msconvert?xcms.exe")}
}
