% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{export_annotations}
\alias{export_annotations}
\title{Export annotations}
\usage{
export_annotations(sqlite_path, excel_path, polarity = "both")
}
\arguments{
\item{sqlite_path}{`character(1)` sqlite path to the annotation results}

\item{excel_path}{`character(1)` path to the excel file to create}

\item{polarity}{`character(1)` polarity to filter annotations, not mandatory}
}
\description{
Export all annotations in a excel file
First sheet will have the annotations regroup by compound
Second will have annotations regroup by ions

Warning ! It export only the annotations with no conflicts !
(Conflicts are when for a group of peaks multiple annotations are possible
(it happens often when for an ion formula refers to multiple compounds))
}
