% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{filter_ms_file}
\alias{filter_ms_file}
\title{Filter MS file}
\usage{
filter_ms_file(ms_file, filter_params)
}
\arguments{
\item{ms_file}{`xcmsRaw` object}

\item{filter_params}{`FilterParam` object}
}
\value{
`xcmsRaw` object filtered
}
\description{
Filter MS file according a rT range in seconds
don't filter on the m/z dimension (too long & too heavy in flash memory)
}
