% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\docType{class}
\name{FilterParam-class}
\alias{FilterParam-class}
\title{A class containing the filter parameters}
\description{
A class containing the filter parameters
}
\section{Slots}{

\describe{
\item{\code{mz_range}}{`numeric(2)` the m/z range to trim all ms files}

\item{\code{rt_range}}{`numeric(2)` the rT range to trim all ms files (in sec)}
}}

