% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{FilterParam}
\alias{FilterParam}
\title{FilterParam}
\usage{
FilterParam(cwt_params, ann_params)
}
\arguments{
\item{cwt_params}{`CentWaveParam`}

\item{ann_params}{`AnnotationParam`}
}
\value{
`FilterParam` object
}
\description{
Construct a FilterParam object
}
