% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_peaks.R
\name{filtrate_ann}
\alias{filtrate_ann}
\title{Filtrate annotation dataframe}
\usage{
filtrate_ann(ann, sigma = 6, perfwhm = 0.6)
}
\arguments{
\item{ann}{the annotation dataframe obtained 
from the function `annotate_peaklists`}

\item{sigma}{ignore}

\item{perfwhm}{ignore}
}
\value{
the annotation dataframe filtered & regrouped
}
\description{
Filtrate annotation dataframe
Foreach compound annotation it will check if 
     the same annotations fall in the same rT
For that it will choose a referenced spectra 
     (the one which as the most isotopologue & 
         the best isotopic score & the less retention time difference)
It will calculate a retention time window which correspond to 
     the rT of the referenced spectra +/- fwhm
It will also regroup lines which correspond 
     to the same annotations but were not grouped by XCMS
}
