% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_chrompeaks.R
\name{find_chrompeaks}
\alias{find_chrompeaks}
\title{Peak picking}
\usage{
find_chrompeaks(ms_file, cwt_params, sample_name)
}
\arguments{
\item{ms_file}{`xcmsRaw` object}

\item{cwt_params}{`CentwaveParam` object}

\item{sample_name}{`character(1)` sample name}
}
\value{
`xcmsSet` object
}
\description{
Execute peak peacking process. It differs from the one from XCMS cause it
take as input an `xcmsRaw` & reconstruct an `xcmsSet` with it
}
