% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_database}
\alias{get_available_database}
\title{Get available databases}
\usage{
get_available_database()
}
\value{
`character` vector with the name of the databases
}
\description{
Get all the names of the available database in package
}
