% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_FUN.R
\name{get_cpd_classes}
\alias{get_cpd_classes}
\title{Get compound classes}
\usage{
get_cpd_classes()
}
\value{
`character vector` compound classes
}
\description{
Get all compound classes recorded in database
}
