% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_FUN.R
\name{get_ions}
\alias{get_ions}
\title{Get ions}
\usage{
get_ions(forms, adduct, instrument)
}
\arguments{
\item{forms}{vector with chemical formulas}

\item{adduct}{subset of the adduct dataframe from enviPat
it needs the columns :
\itemize{
    item Name name of the adduct
    item Mult the multiplicator to apply
    item Charge charge of the adduct
    item Formula_add the chemical formula to add
    item Formula_ded the chemical formula to deduct
}}

\item{instrument}{instrument name for enviPat}
}
\value{
DataFrame with
\itemize{
     \item formula chemical formula
     \item adduct adduct name
     \item ion_formula ion chemical formula
     \item charge charge of the ion
     \item mz m/z
     \item abd relative abundance
     \item iso isotopologue annotation
}
}
\description{
Get all m/z & relative abundance for a set of formulas
}
