% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_FUN.R
\name{get_mz_range}
\alias{get_mz_range}
\title{Get m/z range}
\usage{
get_mz_range(mz, ppm)
}
\arguments{
\item{mz}{`numeric(1)` m/z}

\item{ppm}{`numeric(1)` ppm tolerance}
}
\value{
`numeric(2)` m/z range
}
\description{
Get m/z range with a ppm tolerance
}
