% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_FUN.R
\name{get_mzdev}
\alias{get_mzdev}
\title{Get m/z deviation}
\usage{
get_mzdev(ms_file, mz_range, rt_range)
}
\arguments{
\item{ms_file}{xcmsRaw object}

\item{mz_range}{numeric(2) containing the m/z range to look for}

\item{rt_range}{numeric(2) containing the rt range in sec to look for}
}
\value{
dataframe with the columns:
\itemize{
    \item rt retention time in seconds
    \item int intensity measured
}
}
\description{
Get m/z deviations for an xcmsRaw object
It override the rawMat function from XCMS in order to return rt corrected by
the function `obiwarp`
If the m/z range or rT range is outside from the one from the file it will
return a dataframe with intensities at 0
}
