% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_peaks.R
\name{group_peaks}
\alias{group_peaks}
\title{Group peaklists}
\usage{
group_peaks(ms_files, pd_params, operator, show_pb)
}
\arguments{
\item{ms_files}{XCMSnExp object with peaklists (not checked)}

\item{pd_params}{PeakDensityParam object created by XCMS}

\item{operator}{function to use for parallelization (`\%dopar\%`) 
or not (`\%do\%`)}

\item{show_pb}{boolean indicate if we print a progress bar or not}
}
\value{
XCMSnExp object with the grouped info
}
\description{
Group multiple peaklists from a XCMSnExp object
Use the method group.density from XCMS
The function was modified to parallelize & 
     output the progression on a progress bar
}
\seealso{
xcms::group.density
}
