% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{import_ms_file}
\alias{import_ms_file}
\title{Import ms file}
\usage{
import_ms_file(db, sample_name, raw_file, converter, polarity, filter_params)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{sample_name}{`character(1)` sample name (primary key)}

\item{raw_file}{`character(1)` filepath to the raw file to convert}

\item{converter}{`character(1)` filepath to the msconvert.exe}

\item{polarity}{`character(1)` "positive" or "negative"}

\item{filter_params}{`FilterParam` object}
}
\description{
First it try to convert the file in mzXML
Convert one RAW file to a mzXML file
It will try to convert the file and extract only the scans in positive or
negative
It also trim the file according the rt range and m/z range specified in
the filter_params object with msconvert
Check with XCMS package if the file can be converted
If the file is a CDF it will copy the file instead
If the conversion failed & the file is a mzML or mzXML
it will copy the file instead and try to trim only the rt range
then it record the `xcmsRaw` object obtained in the database if the
conversion was successful
the `xcmsRaw` object is compressed into a `blob` object before inserting in
the database.
This object is recorded in their corresponding "polarity" column (a sample
could contain a `xcmsRaw` in positive AND a `xcmsRaw` in negative !)
Before doing it, we must first use the function `db_record_samples` in order
to prepopulate the database with the sample name (which is the primary key)
}
\seealso{
`convert_file`, `record_ms_file`
}
