% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_FUN.R
\name{load_chem_db}
\alias{load_chem_db}
\title{Load cpd database}
\usage{
load_chem_db(database)
}
\arguments{
\item{database}{`character(1)` database name}
}
\value{
`DataFrame` with columns:
\itemize{
   \item class `character` compound class
   \item name `character` compound name
   \item formula `character` chemical formula
   \item rt `numeric` retention time
}
}
\description{
Load chemical database
}
