% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_FUN.R
\name{load_ion_db}
\alias{load_ion_db}
\title{Get all ions of chemical database}
\usage{
load_ion_db(
  adduct_names,
  instrument,
  database,
  cpd_classes = NULL,
  cpd_names = NULL
)
}
\arguments{
\item{adduct_names}{`character vector`vector of adduct names present in the
enviPat list}

\item{instrument}{`character(1)` instrument name for enviPat}

\item{database}{`character(1)` name of the database to load}

\item{cpd_classes}{`character vector` compound classes to restrict}

\item{cpd_names}{name of compounds to restrict}
}
\value{
DataFrame with
\itemize{
     \item formula chemical formula
     \item class class of the compound
     \item name of the compound
     \item rt retention time of the compound
     \item ion_id id unique for an entry of the database + adduct associated
     \item adduct adduct name
     \item ion_formula ion chemical formula
     \item charge charge of the ion
     \item mz m/z
     \item abd relative abundance
     \item iso isotopologue annotation
}
}
\description{
Get all m/z & relative abundance for the internal database
}
