% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{merge_xsets}
\alias{merge_xsets}
\title{Merge `xcmsSets`}
\usage{
merge_xsets(xset_pos, xset_neg, nsamples)
}
\arguments{
\item{xset_pos}{`xcmSet` in positive mode}

\item{xset_neg}{`xcmSet` in negative mode}

\item{nsamples}{`numeric(1)` number of samples (used only to know where the
column of samples begin in the annotation DataFrame)}
}
\value{
`list` with items:
\itemize{
    \item ann `DataFrame` each line correspond to a compound found
    with the columns:
    \itemize{
        \item group_id `integer` group ID
        \item class `character` cpd class
        \item name `character` name
        \item formula `character` chemical formula
        \item adduct `character` adduct form
        \item ion_formula `character` ion chemical formula
        \item rtdiff `numeric` retention time difference between the measured
         & the expected
        \item rt `numeric` retention time measured meanned accross the
        samples
        \item rtmin `numeric` born min of retention time measured accross the
        samples
        \item rtmax `numeric` born max of the retention time measured accross
        the samples
        \item nsamples `integer` number of samples where the compound was
        found
        \item best_score `numeric` best isotopic score seen
        \item best_deviation_mz `numeric` best m/z deviation seen
        \item best_npeak `integer` best number of isotopologues found
        \item ... `integer` a column for each sample which contain the
        spectra ID
    }
    \item spectras `DataFrame`, each line correspond to a peak annotated with
    its corresponding theoretical peak or the theoretical peak missed,
    with the columns :
    \itemize{
        \item spectra_id `integer` spectra ID
        \item feature_id `integer` feature ID
        \item mz `numeric` m/z
        \item int `numeric` area integrated
        \item abd `numeric` relative abundance
        \item ion_id_theo `integer` ignore
        \item mz_theo `numeric` theoretical m/z
        \item abd_theo `numeric` theoretical relative abundance
        \item iso_theo `character` theoretical isotopologue annotation
    }
    \item spectra_infos `DataFrame`, each line correspond to a spectra
    annotated, with the columns :
    \itemize{
        \item spectra_id `integer` spectra ID
        \item score `numeric` isotopic score observed
        \item deviation_mz `numeric` m/z deviation observed
        \item npeak `integer` number of isotopologue annotated
        \item basepeak_int `numeric` area of the basepeak annotated
        \item sum_int `numeric` cumulative sum off all the area of the
        isotopologues annotated
        \item rt `numeric` retention time
    }
    \item peaks `DataFrame` obtained from XCMS, it contains all peaks peak
    picked, it contains the columns :
    \itemize{
        \item feature_id `integer` feature ID
        \item mz `numeric` m/z
        \item mz_min `numeric` m/z born min
        \item mz_max `numeric` m/z born max
        \item rt `numeric` retention time
        \item rtmin `numeric` retention time born min
        \item rtmax `numeric` retention time born max
        \item int `numeric` area integrated
        \item intb `numeric` area integrated with baseline substracted
        \item maxo `numeric` maximum of intensity
        \item sn `numeric` signal / noise
        \item egauss `numeric` ignore
        \item mu `numeric` ignore
        \item sigma `numeric` ignore,
        \item h `numeric` ignore
        \item f `integer` ID of the ROI in the ROI list constructed by XCMS,
        no use
        \item dppm `numeric` m/z deviation of the peak in ppm
        \item scale `integer` centwave scale used for the integration
        (in scans & not in sec !!!)
        \item scpos `integer` scan position
        \item scmin `integer` scan born min before the optimization of the
        area integrated
        \item scmax `integer` scan born max before the optimization of the
        area integrated
        \item lmin `integer` scan born min after the optimization of the area
        integrated
        \item lmax `integer` scan born max after the optimization of the area
        integrated
        \item sample `character` sample name
        \item polarity `character` polarity
    }
    \item peak_groups `DataFrame`, each line correspond to a group from XCMS,
    the columns are :
    \itemize{
        \item group_id `integer` group ID
        \item polarity `character` polarity
        \item mzmed `numeric` m/z median
        \item mzmin `numeric` m/z born min
        (not the mzmin column from the peaklist !)
        \item mzmax `numeric` m/z born max
        (not the mzmax column from the peaklist !)
        \item rtmed `numeric` rt median
        \item rtmin `numeric` rt born min
        (not the rtmin column from the peaklist !)
        \item rtmax `numeric` rt born max
        (not the rtmax column from the peaklist !)
        \item npeaks `integer` number of peaks grouped
        \item ... `integer` a column for each sample which contain the
        feature ID
    }
}
}
\description{
Merge the `xcmsSet` in positive polarity & the `xcmsSet` in negative polarity
For that it will bind all the annotation results from both & assign new IDs
to distingate if the feature, the group or the spectra was detected
in positive or negative mode
Then it will filtrate the annotation `DataFrame`:
Foreach compound annotation it will check if
     the same annotations fall in the same rT
For that it will choose a referenced spectra
     (the one which as the most adduct forms, then the number of samples
     where founded, then the number of isotopologues, then the less
     retention time difference & the best isotopic)
It will calculate a retention time window which correspond to
     the rT of the referenced spectra +/- fwhm
It will also regroup lines which correspond
     to the same annotations but were not grouped by XCMS
}
