% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{ms_process_polarity}
\alias{ms_process_polarity}
\title{Launch workflow for a polarity}
\usage{
ms_process_polarity(
  sqlite_path,
  samples,
  polarity,
  cwt_params,
  obw_params,
  pd_params,
  ann_params,
  operator = foreach::"\%do\%",
  pb_fct = NULL
)
}
\arguments{
\item{sqlite_path}{`SQLiteConnection`}

\item{samples}{`character vector` sample names in database}

\item{polarity}{`character(1)` muste be "negative" or "positive", used to
load the corresponding `xcmsRaw` & profile matrix}

\item{cwt_params}{`CentwaveParam` object}

\item{obw_params}{`ObiwarpParam` object}

\item{pd_params}{`PeakdensityParam` object}

\item{ann_params}{`AnnotationParam` object}

\item{operator}{`function` to use for parallelization (`\%dopar\%`)
or not (`\%do\%`)}

\item{pb_fct}{`function` used to update the progress bar}
}
\value{
`xcmsSet` with three additional slots :
\itemize{
    \item ann `DataFrame` each line represent an hypothesis annotation
    it contains the columns :
    \itemize{
        \item group_id `integer` group ID
        \item name `character` name
        \item formula `character` chemical formula
        \item adduct `character` adduct form
        \item ion_formula `character` ion chemical formula
        \item rtdiff `numeric` retention time difference between the measured
         & the expected
        \item rt `numeric` retention time measured meanned accross the
        samples
        \item rtmin `numeric` born min of retention time measured accross the
        samples
        \item rtmax `numeric` born max of the retention time measured accross
         the samples
        \item nsamples `integer` number of samples where the compound was
        found
        \item best_score `numeric` best isotopic score seen
        \item best_deviation_mz `numeric` best m/z deviation seen
        \item best_npeak `integer` best number of isotopologues found
        \item ... `integer` a column for each sample which contain the
        spectra ID
    }
    \item spectras `DataFrame`, each line correspond to a peak annotated with
     its corresponding theoretical peak or the theoretical peak missed,
     with the columns :
    \itemize{
        \item spectra_id `integer` spectra ID
        \item feature_id `integer` feature ID
        \item mz `numeric` m/z
        \item int `numeric` area integrated
        \item abd `numeric` relative abundance
        \item ion_id_theo `integer` ignore
        \item mz_theo `numeric` theoretical m/z
        \item abd_theo `numeric` theoretical relative abundance
        \item iso_theo `character` theoretical isotopologue annotation
    }
    \item spectra_infos `DataFrame`, each line correspond to a spectra
    annotated, with the columns :
    \itemize{
        \item spectra_id `integer` spectra ID
        \item score `numeric` isotopic score observed
        \item deviation_mz `numeric` m/z deviation observed
        \item npeak `integer` number of isotopologue annotated
        \item basepeak_int `numeric` area of the basepeak annotated
        \item sum_int `numeric` cumulative sum off all the area of the
        isotopologues annotated
        \item rt `numeric` retention time
    }
}
}
\description{
Launch the workflow for only a unique polarity on xcmsRaws recorded on
a sqlite database
it will launch the workflow which consists in 4 steps :
\itemize{
    \item peak peacking with CentWave algorithm (it will create a list of
    `xcmsSet` objects with the `xcmsRaw` loaded in the database)
    \item alignment with obiwarp which is based on the complete mz-rt data
    \item group peaklists from a `xcmsSet` object using the density method
    \item annotate peaklists from a `xcmsSet`
    it loop through the peaks grouped dataframe
    if the peak match with one of the theoretical monoisotopic from database
        it will search all isotopologue grouped in the rt window of the peak
        +/- fwhm
    it compare then the spectra obtained against the theoretical spectra
        & compute an isotopic score
    The scoring algorithm will search each corresponding observed peak with
        theoreticals. Therefore it contains some important rules :
    \itemize{
         \item an observed peak can only correspond to ONE theoretical peak
          and vice versa
         \item the relative abundance peak must not be under a tolerance
         compared to the theoretical
         but it can be higher since a peak can hide another
         \item the A+x is not searched if the A+x-1 is not found
         (the loop search is stopped)
    }
}
}
