% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obiwarp.R
\name{obiwarp}
\alias{obiwarp}
\title{Correct retention times}
\usage{
obiwarp(
  sqlite_path,
  samples,
  polarity,
  xsets,
  obw_params,
  operator = foreach::"\%do\%",
  pb_fct = NULL
)
}
\arguments{
\item{sqlite_path}{`character(1)` path to the sqlite file}

\item{samples}{`character vector` name of the samples in the database, needed
to load the `xcmsRaw` objects}

\item{polarity}{`character(1)` "negative" or "positive", needed to load the
good `xcmsRaw` objects}

\item{xsets}{`xcmsSet list`}

\item{obw_params}{`ObiwarpParam` object}

\item{operator}{function to use for parallelization (`\%dopar\%`)
or not (`\%do\%`)}

\item{pb_fct}{function used to update the progress bar}
}
\value{
`xcmsSet` which regroup all the files in one object with the rT
corrected
}
\description{
perform retention time correction (alignment) between chromatograms of
     different samples.
the alignment is based on the complete mz-rt data. This method does not
     require any identified peaks or defined features.
This function comes from the XCMS package and only the function `.obiwarp`
     was imported in order to use & export the parallelisation infos
It force to use the sample with the most peak as the center sample
}
\seealso{
xcms::adjustRtime
}
