% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params_to_dataframe,AnnotationParam-method}
\alias{params_to_dataframe,AnnotationParam-method}
\title{Convert `AnnotationParam` to `DataFrame`}
\usage{
\S4method{params_to_dataframe}{AnnotationParam}(object)
}
\arguments{
\item{object}{`AnnotationParam`}
}
\value{
`DataFrame` with one line & the columns:
\itemize{
    \item da_tol `numeric` m/z tolerance in Dalton
    \item rt_tol `numeric` rT tolerance in sec
    \item abd_tol `numeric` relative abundance tolerance, each peak which
    have an higher difference of relative abundance with its corresponding
    theoretical peak will be discarded
    \item adduct_names `character` adduct names from the enviPat package
    collapsed with the character ";"
    \item instrument `character` instrument names from the enviPat package
    \item database `character(1)` name of the database to load
    \item cpd_classes `character` compound classes in database to restrict
    for annotation collapsed with the character ";"
}
}
\description{
Convert a `AnnotationParam` object to a `DataFrame` with one line
}
