% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params_to_dataframe,CameraParam-method}
\alias{params_to_dataframe,CameraParam-method}
\title{Convert `CameraParam` to `DataFrame`}
\usage{
\S4method{params_to_dataframe}{CameraParam}(object)
}
\arguments{
\item{object}{`CameraParam`}
}
\value{
`DataFrame` with one line & the columns:
\itemize{
    \item cores `numeric` number of cores for parallelization
    \item sigma `numeric` multiplier of the standard deviation
    \item perfwhm `numeric` percentage of the FWHM
    \item intval `character` "into", "maxo" or "intb"
    \item cor_eic_th `numeric` correlation threshold
    \item pval `numeric` significant correlation threshold
    \item graphMethod `character` method selection for grouping peaks after
correlation analysis into pseudospectra, could be "hcs" or "lpc"
    \item calcIso `logical` use isotopic relationship for peak grouping
    \item calcCiS `logical` use correlation inside samples for peak grouping
    \item calcCaS `logical` use correlation across samples for peak grouping
    \item maxiso `numeric(1)` max isotopologues
    \item ppm `numeric` ppm tolerance
    \item mzabs `numeric` mDa tolerance
    \item minfrac `numeric` percentage number of samples which must satisfy
     12C/13C rule
    \item max_peaks `numeric` max how much peaks per thread
}
}
\description{
Convert a `CameraParam` object to a `DataFrame` with one line
}
