% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params_to_dataframe,CentWaveParam-method}
\alias{params_to_dataframe,CentWaveParam-method}
\title{Convert `CentWaveParam` to `DataFrame`}
\usage{
\S4method{params_to_dataframe}{CentWaveParam}(object)
}
\arguments{
\item{object}{`CentWaveParam`}
}
\value{
`DataFrame` with one line & the columns:
\itemize{
    \item ppm `numeric` Maximal tolerated m/z deviation in consecutive scans
    in parts sper million (ppm)
    \item peakwidth_min `numeric` Expected approximate peak width min in
    chromatographic space
    \item peakwidth_max `numeric` Expected approximate peak width max in
    chromatographic space
    \item snthresh `numeric` Signal to noise ratio cutoff
    \item prefilter_step `numeric` Mass traces are only retained if they
    contain at least k peaks with intensity >= I
    \item prefilter_level `numeric` Mass traces are only retained if they
    contain at least k peaks with intensity >= I
    \item mzCenterFun `character` Name of the function to calculate the m/z
    center of the chromatographic peak
    \item integrate `integer` Integration method. If unchecked the descent
    is done on the real data, if checked peak limits are found through
    descent on the mexican hat filtered data. Method 1 is very accurate but
    prone to noise, while method 2 is more robust to noise but less exact
    \item mzdiff `numeric` Minimum difference in m/z for peaks with
    overlapping retention times, can be negative to allow overlap
    \item fitgauss `integer` whether or not a Gaussian should be fitted to
    each peak. This affects mostly the retention time position of the peak
    \item noise `numeric` Optional argument which is useful for data that was
    centroided without any intensity threshold, centroids with intensity <
    noise are omitted from ROI detection
    \item verboseColumns `integer` whether additional peak meta data columns
    should be returned, ignore
    \item firstBaselineCheck `integer` Continuous data within regions of
    interest is checked to be above the first baseline
}
}
\description{
Convert a `CentWaveParam` object to a `DataFrame` with one line
}
