% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params_to_dataframe,FilterParam-method}
\alias{params_to_dataframe,FilterParam-method}
\title{Convert `FilterParam` to `DataFrame`}
\usage{
\S4method{params_to_dataframe}{FilterParam}(object)
}
\arguments{
\item{object}{`FilterParam`}
}
\value{
`DataFrame` with one line & the columns:
\itemize{
    \item mz_range_min `numeric` m/z range min
    \item mz_range_max `numeric` m/z range max
    \item rt_range_min `numeric` rT range min
    \item rt_range_max `numeric` rT range max
}
}
\description{
Convert a `FilterParam` object to a `DataFrame` with one line
}
