% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params_to_dataframe,ObiwarpParam-method}
\alias{params_to_dataframe,ObiwarpParam-method}
\title{Convert `ObiwarpParam` to `DataFrame`}
\usage{
\S4method{params_to_dataframe}{ObiwarpParam}(object)
}
\arguments{
\item{object}{`ObiwarpParam` object}
}
\value{
`DataFrame` with one line & the columns:
\itemize{
    \item binSize `numeric` slice of overlapping m/z groups
    \item response `numeric` Defining the responsiveness of warping with
    response = 0 giving linear warping on start and end points and
    response = 100 warping using all bijective anchors
    \item distFun `character` Distance function to be used.
    Allowed values are :
    \itemize{
        \item cor : Pearson's correlation
        \item cor_opt : calculate only 10% diagonal band of distance matrix;
        better runtime)
        \item cov : covariance
        \item prd : product
        \item euc : Euclidian distance
    }
    \item gapInit `numeric` Defining the penalty for gap opening
    \item gapExtend `numeric` Defining the penalty for gap enlargement
    \item factorDiag `numeric` Defining the local weight applied to diagonal
    moves in the alignment
    \item factorGap `numeric` Defining the local weight for gap moves in the
    alignment
    \item localAlignment `integer` Whether a local alignment should be
    performed instead of the default global alignment
    \item initPenalty `numeric` Defining the penalty for initiating an
    alignment (for local alignment only)
}
}
\description{
Convert a `ObiwarpParam` object to a `DataFrame` with one line
}
