% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params_to_dataframe,PeakDensityParam-method}
\alias{params_to_dataframe,PeakDensityParam-method}
\title{Convert `PeakDensityParam` to `DataFrame`}
\usage{
\S4method{params_to_dataframe}{PeakDensityParam}(object)
}
\arguments{
\item{object}{a `PeakDensityParam` object}
}
\value{
`DataFrame` with one line & the columns:
\itemize{
    \item bw `numeric` retention time standard deviation (s) allowed
    \item minFraction `numeric` defining the minimum fraction of samples in
    at least one sample group in which the peaks have to be present to be
    considered as a peak group (feature)
    \item minSamples `integer` with the minimum number of samples in at
    least one sample group in which the peaks have to be detected to be
    considered a peak group (feature)
    \item binSize `numeric` slice of overlapping m/z groups
    \item maxFeatures `integer` with the maximum number of peak groups to be
    identified in a single mz slice
}
}
\description{
Convert a `PeakDensityParam` object to a `DataFrame` with one line
}
