% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_annotation_ms}
\alias{plot_annotation_ms}
\title{Plot MS annotated}
\usage{
plot_annotation_ms(db, name)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{name}{`character(1)` name of the compound}
}
\value{
`plotly` a plotly object
}
\description{
Plot a mass spectrum with all the spectras annotated for a compound
The plot will have in positive the observed ions & in mirror (in negative)
the theoretical ions

It have two JS functions :
\itemize{
    \item when the mouse is over an observed peak we will try to show the
    corresponding theoretical point
    the points are in this order :
    [observed$M, observed$M1, theoretical$M, theoretical$M1]
    so the theoretical must be at the same index but if we start in the
    middle of the array !
    for example for the peak observed M1, the index is 1
    so the theoretical must be at :
        1 + middle = 1 + (length / 2) = 1 + (4 / 2) = 1 + 2 = 3
    \item second is for hiding the annotations bind to the trace and to
    force the relayout between the xaxis range +/- 1
}
}
\examples{
\dontrun{
plot_annotation_ms(db, "LPC 11a:0")
}
}
