% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_composite_ms}
\alias{plot_composite_ms}
\title{Plot Multiple Mass Spectras}
\usage{
plot_composite_ms(spectras)
}
\arguments{
\item{spectras}{`dataframe list` each named item (the names are use to give
a name to the trace) contains a dataframe where each line correspond to a
peak annotated with its corresponding theoretical peak or the theoretical
peak missed, with the columns :
\itemize{
    \item mz `numeric` m/z
    \item int `numeric` area integrated
    \item mz_theo `numeric` theoretical m/z
    \item abd_theo `numeric` theoretical relative abundance
    \item iso_theo `character` theoretical isotopologue annotation
}}
}
\value{
`plotly`
}
\description{
Plot multiple mass spectras on the same plot.
It have in positive the observed ions & in mirror (in negative) the
theoretical ions.
It add two JS functions :
\itemize{
    \item when the mouse is over an observed peak we will try to show the
    corresponding theoretical point
    the points are in this order :
    [observed$M, observed$M1, theoretical$M, theoretical$M1]
    so the theoretical must be at the same index but if we start in the
    middle of the array !
    for example for the peak observed M1, the index is 1
    so the theoretical must be at :
        1 + middle = 1 + (length / 2) = 1 + (4 / 2) = 1 + 2 = 3
    \item second is for hiding the annotations bind to the trace and to
    force the relayout between the xaxis range +/- 1
}
}
