% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_eic_mzdev}
\alias{plot_eic_mzdev}
\title{Plot EIC & m/z deviation}
\usage{
plot_eic_mzdev(db, sample_name, name)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{sample_name}{`character(1)` sample name of the file recorded in db}

\item{name}{`character(1)` compound name}
}
\value{
`plotly`
}
\description{
Plot the EIC for a sample and a compound with the m/z deviations compared to
the theoreticals.
It will trace all the EIC in the same plot for all possible adducts used in
the processing workflow (same for the m/z deviation). The line dashed
correspond to the area not integrated & the line colored the retention time
range where integrated by XCMS. The two line dashed which surround the trace
correspond to the retention time tolerance used for the identification.
This special EIC help to distinguate if this is the peakpicking which turn
wrong or the identification (if it is the rT tolerance or the FWHM window)
It contains a special behavior when the mouse hover a trace : it will display
 all the hovertext of all traces in a unique textbox allowing the user to
  differentiate all the y coordinates of the traces in one shot
It draw also annotation to better view to which adduct correspond the trace
(the annotation is placed at the most intense point for the trace)
A JS function is added to hide the annotation when the trace is hidden
If available it will use the retention time corrected in the slot
`scantime_corrected` added by the function `obiwarp`
}
\examples{
\dontrun{
plot_eic_mzdev(db, "220221CCM_global__01_ssleu_filtered", "LPC 11:0")
}
}
