% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_empty_chromato}
\alias{plot_empty_chromato}
\title{Plot empty chromatogram}
\usage{
plot_empty_chromato(title = "Total Ion Chromatogram(s)")
}
\arguments{
\item{title}{`character(1)` plot title}
}
\value{
`plotly`
}
\description{
Plot an empty chromatogram
It contains a special behavior when the mouse hover a trace : it will display
 all the hovertext of all traces in a unique textbox allowing the user to
  differentiate all the y coordinates of the traces in one shot
}
