% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_empty_MS}
\alias{plot_empty_MS}
\title{Construct empty MS}
\usage{
plot_empty_MS(title = "Mass Spectra", yTitle = "Intensity")
}
\arguments{
\item{title}{`character(1)` title of the plot}

\item{yTitle}{`character(1)` title of the Y axis}
}
\value{
`plotly` object
}
\description{
Construct an empty MS
}
