% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_empty_mzdev}
\alias{plot_empty_mzdev}
\title{Plot empty m/z dev}
\usage{
plot_empty_mzdev(title = "m/z deviation")
}
\arguments{
\item{title}{`character(1)` title}
}
\value{
`plotly`
}
\description{
Plot an empty plot to use for the m/z deviation
}
