% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mzdev}
\alias{plot_mzdev}
\title{Plot m/z deviation}
\usage{
plot_mzdev(db, sample_name, name, adduct_name)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{sample_name}{`character(1)` sample name of the file recorded in db}

\item{name}{`character(1)` compound name}

\item{adduct_name}{`character(1)` name of the adduct used in the
identification step}
}
\value{
`plotly`
}
\description{
Plot the m/z deviation of all the m/z in the file compared to the theoretical
 compound with adduct (only basepeak)
It will trace also a box which represent the m/z tolerance and rT tolernace
used for the identification step
It is useful to use it with the `plot_eic` plot in order to see if there is
a problem during the peak picking or the identification step
If available it will use the retention time corrected in the slot
`scantime_corrected` added by the function `obiwarp`
}
\examples{
\dontrun{
plot_mzdev(
     db,
     "220221CCM_global__01_ssleu_filtered",
     "LPC 11:0",
     "[M+H]+"
)
}
}
