% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{restrict_adducts_polarity}
\alias{restrict_adducts_polarity}
\title{Restrict Annotation parameters}
\usage{
restrict_adducts_polarity(object, polarity)
}
\arguments{
\item{object}{`AnnotationParam`}

\item{polarity}{`character(1)` "positive" or "negative"}
}
\value{
`AnnotationParam` object
}
\description{
will recreate a new `AnnotationParam` with a set of adduct according a
polarity
}
