% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_peaks.R
\name{split_conflicts}
\alias{split_conflicts}
\title{Get conflicts}
\usage{
split_conflicts(ann)
}
\arguments{
\item{ann}{`DataFrame` each line correspond to a compound found
with the columns:
\itemize{
    \item group_id `integer` group ID
    \item class `character` cpd class
    \item name `character` name
    \item formula `character` chemical formula
    \item adduct `character` adduct form
    \item ion_formula `character` ion chemical formula
    \item rtdiff `numeric` retention time difference between the measured &
    the expected
    \item rt `numeric` retention time measured meanned accross the samples
    \item rtmin `numeric` born min of retention time measured accross the
    samples
    \item rtmax `numeric` born max of the retention time measured accross the
     samples
    \item nsamples `integer` number of samples where the compound was found
    \item best_score `numeric` best isotopic score seen
    \item best_deviation_mz `numeric` best m/z deviation seen
    \item best_npeak `integer` best number of isotopologues found
    \item ... `integer` a column for each sample which contain the spectra ID
}}
}
\value{
`list` of two items :
\itemize{
    \item no_conflicts : `DataFrame` each line correspond to a compound found
    with the columns:
    \itemize{
        \item group_id `integer` group ID
        \item class `character` cpd class
        \item name `character` name
        \item formula `character` chemical formula
        \item adduct `character` adduct form
        \item ion_formula `character` ion chemical formula
        \item rtdiff `numeric` retention time difference between the measured
         & the expected
        \item rt `numeric` retention time measured meanned accross the
        samples
        \item rtmin `numeric` born min of retention time measured accross the
        samples
        \item rtmax `numeric` born max of the retention time measured accross
         the samples
        \item nsamples `integer` number of samples where the compound was
        found
        \item best_score `numeric` best isotopic score seen
        \item best_deviation_mz `numeric` best m/z deviation seen
        \item best_npeak `integer` best number of isotopologues found
        \item ... `integer` a column for each sample which contain the
        spectra ID
    }
    \item conflicts : `DataFrame list` each item correspond to the same group
     of peaks where multiple annotations is possible. each dataframe has the
    columns :
    \itemize{
        \item group_id `integer` group ID
        \item class `character` cpd class
        \item name `character` name
        \item formula `character` chemical formula
        \item adduct `character` adduct form
        \item ion_formula `character` ion chemical formula
        \item rtdiff `numeric` retention time difference between the measured
         & the expected
        \item rt `numeric` retention time measured meanned accross the
        samples
        \item rtmin `numeric` born min of retention time measured accross the
        samples
        \item rtmax `numeric` born max of the retention time measured accross
         the samples
        \item nsamples `integer` number of samples where the compound was
        found
        \item best_score `numeric` best isotopic score seen
        \item best_deviation_mz `numeric` best m/z deviation seen
        \item best_npeak `integer` best number of isotopologues found
        \item ... `integer` a column for each sample which contain the
        spectra ID
    }
}
}
\description{
Split the annotation `DataFrame` to get the conflicts
Conflicts are when for a group of peaks multiple annotations are possible
(it happens often when for an ion formula refers to multiple compounds)
}
