% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_peaks.R
\name{summarise_ann}
\alias{summarise_ann}
\title{Summarise annotations}
\usage{
summarise_ann(ann, spectra_infos, nsamples)
}
\arguments{
\item{ann}{`DataFrame` each line correspond to a compound found
with the columns:
\itemize{
    \item group_id `integer` group ID
    \item class `character` cpd class
    \item name `character` name
    \item formula `character` chemical formula
    \item adduct `character` adduct form
    \item ion_formula `character` ion chemical formula
    \item rtdiff `numeric` retention time difference between the measured &
    the expected
    \item rt `numeric` retention time measured meanned accross the samples
    \item rtmin `numeric` born min of retention time measured accross the
    samples
    \item rtmax `numeric` born max of the retention time measured accross the
     samples
    \item nsamples `integer` number of samples where the compound was found
    \item best_score `numeric` best isotopic score seen
    \item best_deviation_mz `numeric` best m/z deviation seen
    \item best_npeak `integer` best number of isotopologues found
    \item ... `integer` a column for each sample which contain the spectra ID
}}

\item{spectra_infos}{`DataFrame`, each line correspond to a spectra
annotated, with the columns :
\itemize{
    \item spectra_id `integer` spectra ID
    \item score `numeric` isotopic score observed
    \item deviation_mz `numeric` m/z deviation observed
    \item npeak `integer` number of isotopologue annotated
    \item basepeak_int `numeric` area of the basepeak annotated
    \item sum_int `numeric` cumulative sum off all the area of the
    isotopologues annotated
    \item rt `numeric` retention time
}}

\item{nsamples}{`integer` the number of samples processed in total, it is use
as an offset on the DataFrame}
}
\value{
`DataFrame` each line represent a compound with the columns
\itemize{
    \item class `character` cpd class
    \item name `character` name of the compound
    \item rt (min) `numeric` meanned rT
    \item Diff rT (sec) `numeric` rT difference between observed &
    theoretical
    \item Adducts `character` all adducts detected separated by a space
    \item nSamples `integer` number of samples where the compound was
    detected
    \item Most intense ion `factor` name of the adduct where the intensity
    measured is the highest
    \item best score (%) `numeric` the highest isotopic score
    \item best m/z dev (mDa) `numeric` the minimal m/z deviation observed
    in mDa
    \item max iso `integer` the highest number of isotopologue for the ions
    reported
    \item ... `integer` a column for each sample which contain the summed
    intensity of ALL basepeaks
}
}
\description{
Summarise the annotations `DataFrame` by compound instead by ion
it will return in the column samples the sum of intensity of ALL basepeaks
}
