% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{AnnotationParam}
\alias{AnnotationParam}
\title{AnnotationParam}
\usage{
AnnotationParam(
  da_tol = 0.015,
  rt_tol = 10,
  abd_tol = 25,
  instrument = "QTOF_XevoG2-S_R25000@200",
  database = "test",
  polarity = "positive",
  cpd_classes = NULL
)
}
\arguments{
\item{da_tol}{`numeric(1)` m/z tolerance in Dalton}

\item{rt_tol}{`numeric(1)` rT tolerance in sec}

\item{abd_tol}{`numeric(1)` relative abundance tolerance, each peak which
have an higher difference of relative abundance with its corresponding
theoretical peak will be discarded}

\item{instrument}{`character(1)` instrument names from the enviPat package}

\item{database}{`character(1)` name of the database to load}

\item{polarity}{`character(1)` "positive" or "negative"}

\item{cpd_classes}{`character vector` compound classes in database to
restrict for annotation}
}
\value{
`AnnotationParam` object
}
\description{
Create an AnnotationParam object
}
\examples{
\dontrun{
AnnotationParam(
     da_tol = .015,
     rt_tol = 10,
     abd_tol = 25,
     instrument = "QTOF_XevoG2-S_R25000@200",
     database = "test"
)}
}
