% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{check_ms_process_args}
\alias{check_ms_process_args}
\title{Check all args of the processing workflow}
\usage{
check_ms_process_args(
  raw_files,
  sqlite_path,
  converter,
  filter_params,
  cwt_params,
  obw_params,
  pd_params,
  ann_params,
  cores
)
}
\arguments{
\item{raw_files}{vector of filepaths to the raw files}

\item{sqlite_path}{filepath to the sqlite database used to save the process
results}

\item{filter_params}{FilterParam object}

\item{cwt_params}{CentWaveParam object created by xcms,}

\item{obw_params}{ObiwarpParam object created by xcms,}

\item{pd_params}{PeakDensityParam object created by xcms,}

\item{ann_params}{AnnotationParam object}

\item{cores}{numeric number of cores to use for the processing workflow
(generaly one core = one file peakpicked)}
}
\description{
Check all arguments of the processing workflow.
     Will raise an error if one argument is incorrect
}
