% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_resolve_conflict}
\alias{db_resolve_conflict}
\title{Resolve annotation conflict}
\usage{
db_resolve_conflict(db, group_id, name)
}
\description{
Resolve an annotation conflict: it will remove all conflicts for a group of
annotations and only keep one
}
