% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{db_resolve_conflict}
\alias{db_resolve_conflict}
\title{Resolve annotation conflict}
\usage{
db_resolve_conflict(db, conflict, i)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{conflict}{`DataFrame` each line correspond to a compound found in the
same group ID with the columns:
\itemize{
    \item group_id `integer` group ID
    \item name `character` name
    \item formula `character` chemical formula
    \item adduct `character` adduct form
    \item ion_formula `character` ion chemical formula
    \item rtdiff `numeric` retention time difference between the measured &
    the expected
    \item rt `numeric` retention time measured meanned accross the samples
    \item rtmin `numeric` born min of retention time measured accross the
    samples
    \item rtmax `numeric` born max of the retention time measured accross the
     samples
    \item nsamples `integer` number of samples where the compound was found
    \item best_score `numeric` best isotopic score seen
    \item best_deviation_mz `numeric` best m/z deviation seen
    \item best_npeak `integer` best number of isotopologues found
    \item ... `integer` a column for each sample which contain the spectra ID
}}

\item{i}{`integer(1)` row ID of the conflict parameter which represent the
annotation to keep}
}
\description{
Resolve an annotation conflict: it will remove all conflicts for a group of
annotations and only keep one
}
