% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_func.R
\name{dbExecute}
\alias{dbExecute}
\title{Send query to sqlite file}
\usage{
dbExecute(db, query, ...)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{query}{`character(1)` query to send}

\item{...}{other parameters passed on the method}
}
\description{
Override the default function `DBI::dbExecute` in order to force the
write process even if the database is locked by another process
}
\seealso{
`DBI::dbExecute`
}
