% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{FilterParam}
\alias{FilterParam}
\title{FilterParam}
\usage{
FilterParam(mz_range = c(300, 1000), rt_range = c(0.7 * 60, 6.3 * 60))
}
\arguments{
\item{mz_range}{`numeric(2)` a m/z range}

\item{rt_range}{`numeric(2)` a rT range in min}
}
\value{
`FilterParam` object
}
\description{
Construct a FilterParam object
}
