% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_peaks.R
\name{group_peaks}
\alias{group_peaks}
\title{Group peaklists}
\usage{
group_peaks(xset, pd_params, operator = foreach::"\%do\%", pb_fct = NULL)
}
\arguments{
\item{xset}{`xcmsSet` object with peaklists}

\item{pd_params}{`PeakDensityParam` object created by XCMS}

\item{operator}{`function` to use for parallelization (`\%dopar\%`)
or not (`\%do\%`)}

\item{pb_fct}{`function` used to update the progress bar}
}
\value{
`xcmsSet` object with the grouped peaks
}
\description{
Group multiple peaklists from a `xcmsSet` object
Use the method group.density from XCMS
The function was modified to parallelize &
     output the progression on a progress bar
}
\seealso{
xcms::group.density
}
