% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot compound heatmap}
\usage{
plot_heatmap(db, names)
}
\arguments{
\item{db}{`SQLiteConnection`}

\item{names}{`character vector` vector of compound names}
}
\value{
`plotly`
}
\description{
Plot a heamap with in x compounds, in y samples and in z their cumulated
intensities of their basepeaks
}
