% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{run_shiny_app}
\alias{run_shiny_app}
\title{Run Shiny app}
\usage{
run_shiny_app(log_file_path = NULL, converter = "pwiz/msconvert.exe")
}
\arguments{
\item{log_file_path}{`character(1)` path to the file were to write the stderr
& stdout}

\item{converter}{`character(1)` path to msconvert.exe}
}
\description{
Run a shiny application in the default web browser
}
\examples{
\dontrun{
     run_shiny_app()
}
}
