% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_peaks.R
\name{summarise_ann}
\alias{summarise_ann}
\title{Resolve annotations conflicts}
\usage{
summarise_ann(ann)
}
\arguments{
\item{ann}{the annotation dataframe obtained 
from the function `annotate_peaklists`}
}
\value{
the annotation dataframe grouped by compound
}
\description{
Summarise the annotations dataframe by compound instead by ion
it will return in the column samples the intensity of 
     the monoisotopic measured for the best adduct form
The best adduct form is the one which was found in the maximum of samples & 
     the minimum of retention time & the best isotopic score
}
